Revert "action cleanup"
This reverts commit 275dddf0a11f30b1ffcf746bddae0be0dfaaeaab.

+++ metta/rl/loss/sl_kickstarter.py
@@ -93,7 +93,7 @@ class SLKickstarter(Loss):
         return Composite(
             # kickstarter loss data
             teacher_action=UnboundedContinuous(
-                shape=(int(self.teacher_policy_spec["action"].shape[0]),), dtype=torch.int32
+                shape=(int(self.teacher_policy_spec["action"].shape[0]),), dtype=torch.int64
             ),
             teacher_value=scalar_f32,
         )
@@ -109,13 +109,13 @@ class SLKickstarter(Loss):
         # Teacher forward pass
         teacher_td = policy_td.select(*self.teacher_policy_spec.keys(include_nested=True)).clone()
         teacher_td = self.teacher_policy(teacher_td, action=None)
-        teacher_action = teacher_td["action"].to(dtype=torch.int32).detach()
+        teacher_action = teacher_td["action"].to(dtype=torch.int64).detach()
         teacher_value = teacher_td["values"].to(dtype=torch.float32).detach()
 
         # Student forward pass
         student_td = policy_td.select(*self.policy_experience_spec.keys(include_nested=True)).clone()
         student_td = self.policy(student_td, action=None)
-        student_action = student_td["action"].to(dtype=torch.int32)
+        student_action = student_td["action"].to(dtype=torch.int64)
         student_value = student_td["values"].to(dtype=torch.float32)
 
         # Calculate annealing coefficient
